--- src/application/hydratorManager.js
@@ -1,8 +1,0 @@
-const {hydrateHorizontalData, hydratePixelData} = require('../domain/pixelsHydration');
-
-module.exports = function () {
-    return {
-        hydrateHorizontalData,
-        hydratePixelData,
-    };
-};
