+++ src/application/hydratorManager.mjs
@@ -1,0 +1,8 @@
+import {hydrateHorizontalData, hydratePixelData} from '../domain/pixelsHydration.mjs';
+
+export default function () {
+    return {
+        hydrateHorizontalData,
+        hydratePixelData,
+    };
+}
