--- src/application/parserManager.js
@@ -1,5 +1,0 @@
-const builder = require('../domain/parser');
-
-module.exports = function (options) {
-    return builder(options);
-};
