+++ test/fixtures/application/parser/code.js
@@ -1,0 +1,18 @@
+
+function lospepeh() {
+    const pepe = 1;
+    const pepa = 2;
+    const pepo = [1, 2, 3, 4, 5, 6];
+
+    const pepi = pepo.map(piper => {
+        if (piper === pepe) {
+            console.log('pepe');
+        } else if (piper === pepa) {
+            console.log('pepa');
+        } else {
+            console.log('pepo');
+        }
+
+        return piper + 5;
+    });
+}
