+++ test/hydrationManager.test.mjs
@@ -1,7 +1,7 @@
-const path = require('path');
-const fs = require('fs-extra');
+import path from 'path';
+import fs from 'fs-extra';
 
-const hydratorManager = require('../src/application/hydratorManager');
+import {hydrateHorizontalData} from '../src/application/hydratorManager';
 
 describe('File Manager', () => {
 
@@ -9,7 +9,6 @@ describe('File Manager', () => {
         // Given
         const input = (await fs.readFile(path.resolve('test/fixtures/expectedOutputFromNormalizedData.txt'), 'utf8')).toString();
         const expectedResult = (await fs.readFile(path.resolve('test/fixtures/expectedOutputFromHydratedData.txt'), 'utf8'));
-        const {hydrateHorizontalData} = hydratorManager();
 
         // When
         const result = await hydrateHorizontalData(input);
