+++ test/main.e2e.test.mjs
@@ -1,11 +1,12 @@
-const path = require('path');
+import {jest} from '@jest/globals';
+import path from 'path';
 
-const fs = require('fs-extra');
-const axios = require('axios');
+import fs from 'fs-extra';
+import axios from 'axios';
 
 jest.mock('axios');
 
-const execute = require('../src/index');
+import execute from '../src/index.mjs';
 
 describe('Test execution interfaces', () => {
